requirements.txt
fastapi==0.108.0
pytest==7.1.3
uvicorn==0.20.0
sqlalchemy==2.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pre-commit
pydantic==1.10.13